change u32 to u64

+++ core/src/models/animation.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 pub struct Animation {
     pub file_id: String,
     pub file_unique_id: String,
-    pub width: u32,
-    pub height: u32,
-    pub duration: u32,
+    pub width: u64,
+    pub height: u64,
+    pub duration: u64,
     pub thumbnail: Option<PhotoSize>,
     pub file_name: Option<String>,
     pub mime_type: Option<String>,
