+++ core/src/models/location.rs
@@ -5,7 +5,7 @@ pub struct Location {
     pub longitude: f64,
     pub latitude: f64,
     pub horizontal_accuracy: Option<f64>,
-    pub live_period: Option<u32>,
+    pub live_period: Option<u64>,
     pub heading: Option<u16>,
-    pub proximity_alert_radius: Option<u32>,
+    pub proximity_alert_radius: Option<u64>,
 }
