+++ core/src/models/successful_payment.rs
@@ -5,7 +5,7 @@ use crate::models::order_info::OrderInfo;
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct SuccessfulPayment {
     pub currency: String,
-    pub total_amount: u32,
+    pub total_amount: u64,
     pub invoice_payload: String,
     pub shipping_option_id: Option<String>,
     pub order_info: Option<OrderInfo>,
