+++ core/src/models/video.rs
@@ -6,9 +6,9 @@ use crate::models::photo_size::PhotoSize;
 pub struct Video {
     pub file_id: String,
     pub file_unique_id: String,
-    pub width: u32,
-    pub height: u32,
-    pub duration: u32,
+    pub width: u64,
+    pub height: u64,
+    pub duration: u64,
     pub thumbnail: Option<PhotoSize>,
     pub file_name: Option<String>,
     pub mime_type: Option<String>,
