+++ core/src/models/voice.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 pub struct Voice {
     pub file_id: String,
     pub file_unique_id: String,
-    pub duration: u32,
+    pub duration: u64,
     pub mime_type: Option<String>,
     pub file_size: Option<u64>,
 }
