+++ core/src/params/updates_params.rs
@@ -8,9 +8,9 @@ pub struct GetUpdatesParams {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub offset: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<u32>,
+    pub limit: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub timeout: Option<u32>,
+    pub timeout: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub allowed_updates: Option<Vec<AllowedUpdate>>,
 }
